Renderer is now a base class and not a concrete implementation. RendererGL20 class used as a reference implementation. Will add RendererGL10 soon. Example project for Bill Gates 98 jdolan review pls
