+++ ObjectivelyMVC.vs15/Example.vcxproj.filters
@@ -1,0 +1,10 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <ItemGroup>
+    <ClCompile Include="..\Examples\Hello.c" />
+    <ClCompile Include="..\Examples\HelloViewController.c" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="..\Examples\HelloViewController.h" />
+  </ItemGroup>
+</Project>
