+++ Sources/ObjectivelyMVC/Renderer.c
@@ -28,38 +28,6 @@
 #include <ObjectivelyMVC/Renderer.h>
 #include <ObjectivelyMVC/View.h>
 
-#pragma mark - OpenGL entry points
-
-static PFNGLBINDBUFFERPROC glBindBuffer;
-static PFNGLBUFFERDATAPROC glBufferData;
-static PFNGLBUFFERSUBDATAPROC glBufferSubData;
-static PFNGLDELETEBUFFERSPROC glDeleteBuffers;
-static PFNGLDISABLEVERTEXATTRIBARRAYPROC glDisableVertexAttribArray;
-static PFNGLENABLEVERTEXATTRIBARRAYPROC glEnableVertexAttribArray;
-static PFNGLGENBUFFERSPROC glGenBuffers;
-static PFNGLUNIFORM4FVPROC glUniform4fv;
-static PFNGLUSEPROGRAMPROC glUseProgram;
-static PFNGLVERTEXATTRIBPOINTERPROC glVertexAttribPointer;
-
-#pragma mark - Image data
-
-static Image *_null;
-static const unsigned char _nullData[] = {
-	0x89, 0x50, 0x4e, 0x47, 0x0d, 0x0a, 0x1a, 0x0a, 0x00, 0x00, 0x00, 0x0d,
-	0x49, 0x48, 0x44, 0x52, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x01,
-	0x08, 0x02, 0x00, 0x00, 0x00, 0x90, 0x77, 0x53, 0xde, 0x00, 0x00, 0x00,
-	0x09, 0x70, 0x48, 0x59, 0x73, 0x00, 0x00, 0x0b, 0x13, 0x00, 0x00, 0x0b,
-	0x13, 0x01, 0x00, 0x9a, 0x9c, 0x18, 0x00, 0x00, 0x00, 0x07, 0x74, 0x49,
-	0x4d, 0x45, 0x07, 0xe0, 0x0a, 0x10, 0x14, 0x18, 0x17, 0xda, 0xde, 0xe8,
-	0x2e, 0x00, 0x00, 0x00, 0x1d, 0x69, 0x54, 0x58, 0x74, 0x43, 0x6f, 0x6d,
-	0x6d, 0x65, 0x6e, 0x74, 0x00, 0x00, 0x00, 0x00, 0x00, 0x43, 0x72, 0x65,
-	0x61, 0x74, 0x65, 0x64, 0x20, 0x77, 0x69, 0x74, 0x68, 0x20, 0x47, 0x49,
-	0x4d, 0x50, 0x64, 0x2e, 0x65, 0x07, 0x00, 0x00, 0x00, 0x0c, 0x49, 0x44,
-	0x41, 0x54, 0x08, 0xd7, 0x63, 0xf8, 0xff, 0xff, 0x3f, 0x00, 0x05, 0xfe,
-	0x02, 0xfe, 0xdc, 0xcc, 0x59, 0xe7, 0x00, 0x00, 0x00, 0x00, 0x49, 0x45,
-	0x4e, 0x44, 0xae, 0x42, 0x60, 0x82
-};
-
 #define _Class _Renderer
 
 #pragma mark - Object
@@ -71,12 +39,6 @@ static void dealloc(Object *self) {
 
 	Renderer *this = (Renderer *) self;
 
-	glDeleteTextures(1, &this->nullTexture);
-
-	glDeleteBuffers(1, &this->texcoordBuffer);
-	glDeleteBuffers(1, &this->vertexBuffer);
-
-	release(this->program);
 	release(this->views);
 
 	super(Object, self, dealloc);
@@ -98,18 +60,6 @@ static void addView(Renderer *self, View *view) {
  */
 static void beginFrame(Renderer *self) {
 
-	glEnable(GL_BLEND);
-	glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
-
-	glEnable(GL_SCISSOR_TEST);
-
-	glBindTexture(GL_TEXTURE_2D, self->nullTexture);
-
-	glUseProgram(self->program->name);
-
-	glEnableVertexAttribArray(self->vertexAttribute);
-	glEnableVertexAttribArray(self->texcoordAttribute);
-
 	$(self, setDrawColor, &Colors.White);
 }
 
@@ -121,36 +71,15 @@ static ident createTexture(const Renderer *self, const SDL_Surface *surface) {
 
 	assert(surface);
 
-	GLenum format;
-	switch (surface->format->BytesPerPixel) {
-		case 1:
-			format = GL_LUMINANCE;
-			break;
-		case 3:
-			format = GL_RGB;
-			break;
-		case 4:
-			format = GL_RGBA;
-			break;
-		default:
-			MVC_LogError("Invalid surface format: %s\n",
-					 SDL_GetPixelFormatName(surface->format->format));
-			return -1;
-	}
+	return NULL;
+}
 
-	GLuint texture;
-	glGenTextures(1, &texture);
+/**
+ * @fn void Renderer::destroyTexture(const Renderer *self, ident texture)
+ * @memberof Renderer
+ */
+static void destroyTexture(const Renderer *self, ident texture) {
 
-	glBindTexture(GL_TEXTURE_2D, texture);
-
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-	glTexParameteri(GL_TEXTURE_2D, GL_GENERATE_MIPMAP, GL_FALSE);
-
-	glTexImage2D(GL_TEXTURE_2D, 0, format, surface->w, surface->h, 0, format,
-				 GL_UNSIGNED_BYTE, surface->pixels);
-
-	return (ident)texture;
 }
 
 /**
@@ -171,12 +100,6 @@ static void drawLine(const Renderer *self, const SDL_Point *points) {
 static void drawLines(const Renderer *self, const SDL_Point *points, size_t count) {
 
 	assert(points);
-
-	glBindBuffer(GL_ARRAY_BUFFER, self->vertexBuffer);
-	glBufferSubData(GL_ARRAY_BUFFER, 0, count * sizeof(SDL_Point), points);
-	glBindBuffer(GL_ARRAY_BUFFER, 0);
-
-	glDrawArrays(GL_LINE_STRIP, 0, count);
 }
 
 /**
@@ -186,26 +109,6 @@ static void drawLines(const Renderer *self, const SDL_Point *points, size_t coun
 static void drawRect(const Renderer *self, const SDL_Rect *rect) {
 
 	assert(rect);
-
-	GLint verts[8];
-
-	verts[0] = rect->x;
-	verts[1] = rect->y;
-
-	verts[2] = rect->x + rect->w;
-	verts[3] = rect->y;
-
-	verts[4] = rect->x + rect->w;
-	verts[5] = rect->y + rect->h;
-
-	verts[6] = rect->x;
-	verts[7] = rect->y + rect->h;
-
-	glBindBuffer(GL_ARRAY_BUFFER, self->vertexBuffer);
-	glBufferSubData(GL_ARRAY_BUFFER, 0, sizeof(verts), verts);
-	glBindBuffer(GL_ARRAY_BUFFER, 0);
-
-	glDrawArrays(GL_LINE_LOOP, 0, 4);
 }
 
 /**
@@ -215,26 +118,6 @@ static void drawRect(const Renderer *self, const SDL_Rect *rect) {
 static void drawRectFilled(const Renderer *self, const SDL_Rect *rect) {
 
 	assert(rect);
-
-	GLint verts[8];
-
-	verts[0] = rect->x - 1;
-	verts[1] = rect->y;
-
-	verts[2] = rect->x + rect->w;
-	verts[3] = rect->y;
-
-	verts[4] = rect->x + rect->w;
-	verts[5] = rect->y + rect->h + 1;
-
-	verts[6] = rect->x - 1;
-	verts[7] = rect->y + rect->h + 1;
-
-	glBindBuffer(GL_ARRAY_BUFFER, self->vertexBuffer);
-	glBufferSubData(GL_ARRAY_BUFFER, 0, sizeof(verts), verts);
-	glBindBuffer(GL_ARRAY_BUFFER, 0);
-
-	glDrawArrays(GL_TRIANGLE_FAN, 0, 4);
 }
 
 /**
@@ -244,23 +127,6 @@ static void drawRectFilled(const Renderer *self, const SDL_Rect *rect) {
 static void drawTexture(const Renderer *self, ident texture, const SDL_Rect *rect) {
 
 	assert(rect);
-
-	const GLfloat texcoords[] = {
-		0.0, 0.0,
-		1.0, 0.0,
-		1.0, 1.0,
-		0.0, 1.0
-	};
-
-	glBindTexture(GL_TEXTURE_2D, (GLuint)texture);
-
-	glBindBuffer(GL_ARRAY_BUFFER, self->texcoordBuffer);
-	glBufferSubData(GL_ARRAY_BUFFER, 0, sizeof(texcoords), texcoords);
-	glBindBuffer(GL_ARRAY_BUFFER, 0);
-
-	$(self, drawRectFilled, rect);
-
-	glBindTexture(GL_TEXTURE_2D, self->nullTexture);
 }
 
 /**
@@ -269,27 +135,6 @@ static void drawTexture(const Renderer *self, ident texture, const SDL_Rect *rec
  */
 static void endFrame(Renderer *self) {
 
-	glDisableVertexAttribArray(self->texcoordAttribute);
-	glDisableVertexAttribArray(self->vertexAttribute);
-
-	glUseProgram(0);
-
-	SDL_Window *window = SDL_GL_GetCurrentWindow();
-
-	int dw, dh;
-	SDL_GL_GetDrawableSize(window, &dw, &dh);
-
-	glScissor(0, 0, dw, dh);
-
-	glDisable(GL_SCISSOR_TEST);
-
-	glBlendFunc(GL_ONE, GL_ZERO);
-	glDisable(GL_BLEND);
-
-	const GLenum err = glGetError();
-	if (err) {
-		MVC_LogError("GL error: %d\n", err);
-	}
 }
 
 /**
@@ -327,15 +172,20 @@ static Order render_sort(const ident a, const ident b) {
 static void render_renderView(const Array *array, ident obj, ident data) {
 
 	View *view = (View *) obj;
+	Renderer *renderer = (Renderer *) data;
 
 	const SDL_Rect frame = $(view, clippingFrame);
 	if (frame.w && frame.h) {
 
-		const SDL_Rect scissor = MVC_TransformToWindow($(view, window), &frame);
+		SDL_Rect scissor = MVC_TransformToWindow($(view, window), &frame);
+		scissor.x--;
+		scissor.y--;
+		scissor.w++;
+		scissor.h++;
 
-		glScissor(scissor.x - 1, scissor.y - 1, scissor.w + 1, scissor.h + 1);
+		$(renderer, setScissor, &scissor);
 
-		$(view, render, (Renderer *) data);
+		$(view, render, renderer);
 	}
 }
 
@@ -358,56 +208,6 @@ static void render(Renderer *self) {
  */
 static void renderDeviceDidReset(Renderer *self) {
 
-	glGenBuffers(1, &self->vertexBuffer);
-	glBindBuffer(GL_ARRAY_BUFFER, self->vertexBuffer);
-	glBufferData(GL_ARRAY_BUFFER, 0x10000, NULL, GL_DYNAMIC_DRAW);
-	glBindBuffer(GL_ARRAY_BUFFER, 0);
-
-	glGenBuffers(1, &self->texcoordBuffer);
-	glBindBuffer(GL_ARRAY_BUFFER, self->texcoordBuffer);
-	glBufferData(GL_ARRAY_BUFFER, 0x10000, NULL, GL_DYNAMIC_DRAW);
-	glBindBuffer(GL_ARRAY_BUFFER, 0);
-
-	self->nullTexture = $(self, createTexture, _null->surface);
-
-	self->program = (Program *) $(alloc(Program), init);
-	assert(self->program);
-
-	$(self->program, attachShaderSource, GL_VERTEX_SHADER, "\
-		#version 120 \n\
-		attribute vec2 vertex; \
-		attribute vec2 texcoord; \
-		varying vec2 st; \
-		void main() { \
-			gl_Position = gl_ProjectionMatrix * vec4(vertex, 0.0, 1.0); \
-			st = texcoord; \
-		}");
-
-	$(self->program, attachShaderSource, GL_FRAGMENT_SHADER, "\
-		#version 120 \n\
-		uniform vec4 color; \
-		uniform sampler2D sampler; \
-		varying vec2 st; \
-		void main() { \
-			gl_FragColor = color * texture2D(sampler, st); \
-		}");
-
-	$(self->program, link);
-	assert(self->program->name);
-
-	self->vertexAttribute = $(self->program, getAttributeLocation, "vertex");
-	self->texcoordAttribute = $(self->program, getAttributeLocation, "texcoord");
-	self->colorUniform = $(self->program, getUniformLocation, "color");
-
-	glBindBuffer(GL_ARRAY_BUFFER, self->vertexBuffer);
-	glVertexAttribPointer(self->vertexAttribute, 2, GL_INT, GL_FALSE, 0, NULL);
-	glBindBuffer(GL_ARRAY_BUFFER, 0);
-
-	glBindBuffer(GL_ARRAY_BUFFER, self->texcoordBuffer);
-	glVertexAttribPointer(self->texcoordAttribute, 2, GL_FLOAT, GL_FALSE, 0, NULL);
-	glBindBuffer(GL_ARRAY_BUFFER, 0);
-
-	assert(glGetError() == GL_NO_ERROR);
 }
 
 /**
@@ -415,17 +215,18 @@ static void renderDeviceDidReset(Renderer *self) {
  * @memberof Renderer
  */
 static void setDrawColor(Renderer *self, const SDL_Color *color) {
-	glUniform4fv(self->colorUniform, 1, (const GLfloat []) {
-		color->r / 255.0,
-		color->g / 255.0,
-		color->b / 255.0,
-		color->a / 255.0
-	});
+
+}
+
+/**
+ * @fn void Renderer::setScissor(Renderer *self, const SDL_Rect *rect)
+ * @memberof Renderer
+ */
+static void setScissor(Renderer *self, const SDL_Rect *rect) {
+
 }
 
 #pragma mark - Class lifecycle
-
-#include <ObjectivelyMVC/Program.h>
 
 /**
  * @see Class::initialize(Class *)
@@ -437,6 +238,7 @@ static void initialize(Class *clazz) {
 	((RendererInterface *) clazz->def->interface)->addView = addView;
 	((RendererInterface *) clazz->def->interface)->beginFrame = beginFrame;
 	((RendererInterface *) clazz->def->interface)->createTexture = createTexture;
+	((RendererInterface *) clazz->def->interface)->destroyTexture = destroyTexture;
 	((RendererInterface *) clazz->def->interface)->drawLine = drawLine;
 	((RendererInterface *) clazz->def->interface)->drawLines = drawLines;
 	((RendererInterface *) clazz->def->interface)->drawRect = drawRect;
@@ -447,26 +249,7 @@ static void initialize(Class *clazz) {
 	((RendererInterface *) clazz->def->interface)->render = render;
 	((RendererInterface *) clazz->def->interface)->renderDeviceDidReset = renderDeviceDidReset;
 	((RendererInterface *) clazz->def->interface)->setDrawColor = setDrawColor;
-
-	_null = $(alloc(Image), initWithBytes, _nullData, lengthof(_nullData));
-
-	MVC_GL_GetProcAddress(glBindBuffer);
-	MVC_GL_GetProcAddress(glBufferData);
-	MVC_GL_GetProcAddress(glBufferSubData);
-	MVC_GL_GetProcAddress(glDeleteBuffers);
-	MVC_GL_GetProcAddress(glDisableVertexAttribArray);
-	MVC_GL_GetProcAddress(glEnableVertexAttribArray);
-	MVC_GL_GetProcAddress(glGenBuffers);
-	MVC_GL_GetProcAddress(glUniform4fv);
-	MVC_GL_GetProcAddress(glUseProgram);
-	MVC_GL_GetProcAddress(glVertexAttribPointer);
-}
-
-/**
- * @see Class::destroy(Class *)
- */
-static void destroy(Class *clazz) {
-	release(_null);
+	((RendererInterface *) clazz->def->interface)->setScissor = setScissor;
 }
 
 Class _Renderer = {
@@ -475,8 +258,7 @@ Class _Renderer = {
 	.instanceSize = sizeof(Renderer),
 	.interfaceOffset = offsetof(Renderer, interface),
 	.interfaceSize = sizeof(RendererInterface),
-	.initialize = initialize,
-	.destroy = destroy,
+	.initialize = initialize
 };
 
 #undef _Class
